huggingface_deploy_llm/requirements.txt
llama-cpp-python
langchain
fastapi